moved assertions into assertions module
